+++ e2e/aws-lambda/.aws-sam/build.toml
@@ -1,0 +1,12 @@
+# This file is auto generated by SAM CLI build command
+
+[function_build_definitions.30fd04b9-04cc-44af-acf1-bcfdc8455038]
+codeuri = "/Users/vkorolik/work/highlight/e2e/aws-lambda"
+runtime = "nodejs22.x"
+architecture = "x86_64"
+handler = "src/handlers/api.handler"
+manifest_hash = "5087545de796328464e8d8fe92e4ffd4"
+packagetype = "Zip"
+functions = ["apiFunction"]
+
+[layer_build_definitions]
