--- e2e/aws-lambda/env.json
@@ -1,6 +1,0 @@
-{
-	"ApiFunction": { "HIGHLIGHT_PROJECT_ID": "your-project-id" },
-	"ErrorFunction": { "HIGHLIGHT_PROJECT_ID": "your-project-id" },
-	"MetricsFunction": { "HIGHLIGHT_PROJECT_ID": "your-project-id" },
-	"ProcessorFunction": { "HIGHLIGHT_PROJECT_ID": "your-project-id" }
-}
