--- e2e/aws-lambda/events/api-event.json
@@ -1,66 +1,0 @@
-{
-	"resource": "/api",
-	"path": "/api",
-	"httpMethod": "GET",
-	"headers": {
-		"Accept": "application/json",
-		"Accept-Encoding": "gzip, deflate, br",
-		"Host": "api.example.com",
-		"User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:91.0) Gecko/20100101 Firefox/91.0",
-		"X-Forwarded-For": "127.0.0.1",
-		"X-Session-Id": "test-session-abc123"
-	},
-	"multiValueHeaders": {
-		"Accept": ["application/json"],
-		"Accept-Encoding": ["gzip, deflate, br"],
-		"Host": ["api.example.com"],
-		"User-Agent": [
-			"Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:91.0) Gecko/20100101 Firefox/91.0"
-		],
-		"X-Forwarded-For": ["127.0.0.1"],
-		"X-Session-Id": ["test-session-abc123"]
-	},
-	"queryStringParameters": {
-		"param1": "value1",
-		"param2": "value2"
-	},
-	"multiValueQueryStringParameters": {
-		"param1": ["value1"],
-		"param2": ["value2"]
-	},
-	"pathParameters": null,
-	"stageVariables": null,
-	"requestContext": {
-		"resourceId": "123456",
-		"resourcePath": "/api",
-		"httpMethod": "GET",
-		"extendedRequestId": "request-id-1",
-		"requestTime": "09/Apr/2023:12:34:56 +0000",
-		"path": "/prod/api",
-		"accountId": "123456789012",
-		"protocol": "HTTP/1.1",
-		"stage": "prod",
-		"domainPrefix": "api",
-		"requestTimeEpoch": 1609459200000,
-		"requestId": "c6af9ac6-7b61-11e6-9a41-93e8deadbeef",
-		"identity": {
-			"cognitoIdentityPoolId": null,
-			"accountId": null,
-			"cognitoIdentityId": null,
-			"caller": null,
-			"sourceIp": "127.0.0.1",
-			"principalOrgId": null,
-			"accessKey": null,
-			"cognitoAuthenticationType": null,
-			"cognitoAuthenticationProvider": null,
-			"userArn": null,
-			"userAgent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.15; rv:91.0) Gecko/20100101 Firefox/91.0",
-			"user": null,
-			"sessionId": "test-session-abc123"
-		},
-		"domainName": "api.example.com",
-		"apiId": "api-id"
-	},
-	"body": null,
-	"isBase64Encoded": false
-}
