+++ e2e/aws-lambda/events/event-get-all-items.json
@@ -1,0 +1,3 @@
+{
+	"httpMethod": "GET"
+}
