+++ e2e/aws-lambda/events/event-get-by-id.json
@@ -1,0 +1,6 @@
+{
+	"httpMethod": "GET",
+	"pathParameters": {
+		"id": "id1"
+	}
+}
