+++ e2e/aws-lambda/events/event-post-item.json
@@ -1,0 +1,4 @@
+{
+	"httpMethod": "POST",
+	"body": "{\"id\": \"id1\",\"name\": \"name1\"}"
+}
