--- e2e/aws-lambda/events/sqs-event.json
@@ -1,79 +1,0 @@
-{
-	"Records": [
-		{
-			"messageId": "19dd0b57-b21e-4ac1-bd88-01bbb068cb78",
-			"receiptHandle": "MessageReceiptHandle",
-			"body": "{\"type\":\"notification\",\"content\":\"This is a test notification\"}",
-			"attributes": {
-				"ApproximateReceiveCount": "1",
-				"SentTimestamp": "1523232000000",
-				"SenderId": "123456789012",
-				"ApproximateFirstReceiveTimestamp": "1523232000001"
-			},
-			"messageAttributes": {
-				"secureSessionId": {
-					"stringValue": "example-session-123",
-					"dataType": "String"
-				},
-				"requestId": {
-					"stringValue": "req-123456789",
-					"dataType": "String"
-				}
-			},
-			"md5OfBody": "7b270e59b47ff90a553787216d55d91d",
-			"eventSource": "aws:sqs",
-			"eventSourceARN": "arn:aws:sqs:us-west-2:123456789012:example-queue",
-			"awsRegion": "us-west-2"
-		},
-		{
-			"messageId": "87a9126c-5d32-4a21-9e3d-03a1f4c6e624",
-			"receiptHandle": "MessageReceiptHandle2",
-			"body": "{\"type\":\"order\",\"content\":\"This is a test order\",\"orderId\":\"ord-12345\"}",
-			"attributes": {
-				"ApproximateReceiveCount": "1",
-				"SentTimestamp": "1523232000000",
-				"SenderId": "123456789012",
-				"ApproximateFirstReceiveTimestamp": "1523232000001"
-			},
-			"messageAttributes": {
-				"secureSessionId": {
-					"stringValue": "example-session-456",
-					"dataType": "String"
-				},
-				"requestId": {
-					"stringValue": "req-987654321",
-					"dataType": "String"
-				}
-			},
-			"md5OfBody": "a0d8def2885c8e34e3c755e2ab549864",
-			"eventSource": "aws:sqs",
-			"eventSourceARN": "arn:aws:sqs:us-west-2:123456789012:example-queue",
-			"awsRegion": "us-west-2"
-		},
-		{
-			"messageId": "634c94a0-cefd-4af9-a2b5-0e6196516c5a",
-			"receiptHandle": "MessageReceiptHandle3",
-			"body": "{\"type\":\"data_sync\",\"content\":\"This is a test data sync\",\"dataSource\":\"CRM\"}",
-			"attributes": {
-				"ApproximateReceiveCount": "1",
-				"SentTimestamp": "1523232000000",
-				"SenderId": "123456789012",
-				"ApproximateFirstReceiveTimestamp": "1523232000001"
-			},
-			"messageAttributes": {
-				"secureSessionId": {
-					"stringValue": "example-session-789",
-					"dataType": "String"
-				},
-				"requestId": {
-					"stringValue": "req-abcdefg",
-					"dataType": "String"
-				}
-			},
-			"md5OfBody": "c6e87d08e25eaf13a54a3c2aa927a73f",
-			"eventSource": "aws:sqs",
-			"eventSourceARN": "arn:aws:sqs:us-west-2:123456789012:example-queue",
-			"awsRegion": "us-west-2"
-		}
-	]
-}
