+++ e2e/aws-lambda/template.yaml
@@ -1,104 +1,52 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: AWS::Serverless-2016-10-31
-Description: >
-    Highlight Node.js SDK AWS Lambda Example
-    Sample application showcasing Highlight SDK integration with AWS Lambda
+AWSTemplateFormatVersion: 2010-09-09
+Description: >-
+    aws-lambda
+Transform:
+    - AWS::Serverless-2016-10-31
 
-Parameters:
-    HighlightProjectID:
-        Type: String
-        Description: Your Highlight Project ID
-        Default: 'your-project-id'
-
-Globals:
-    Function:
-        Timeout: 30
-        MemorySize: 512
-        Runtime: nodejs18.x
-        Architectures:
-            - x86_64
-        Environment:
-            Variables:
-                HIGHLIGHT_PROJECT_ID: !Ref HighlightProjectID
-
+# Resources declares the AWS resources that you want to include in the stack
+# https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/resources-section-structure.html
 Resources:
-    # Lambda function handling HTTP events
-    ApiFunction:
+    # Each Lambda function is defined by properties:
+    # https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
+
+    # This is a Lambda function config associated with the source code: get-all-items.js
+    apiFunction:
         Type: AWS::Serverless::Function
         Properties:
-            CodeUri: ./
-            Handler: src/api.handler
+            Handler: src/handlers/api.handler
+            Description: A simple example includes a HTTP get method
             Events:
                 GetApi:
                     Type: Api
                     Properties:
-                        Path: /api
+                        Path: /
                         Method: GET
                 PostApi:
                     Type: Api
                     Properties:
-                        Path: /api
+                        Path: /
                         Method: POST
-
-    # Lambda function for processing events
-    ProcessorFunction:
-        Type: AWS::Serverless::Function
-        Properties:
-            CodeUri: ./
-            Handler: src/processor.handler
-            Events:
-                SQSEvent:
-                    Type: SQS
-                    Properties:
-                        Queue: !GetAtt ProcessingQueue.Arn
-                        BatchSize: 10
-
-    # Lambda function for error demonstration
-    ErrorFunction:
-        Type: AWS::Serverless::Function
-        Properties:
-            CodeUri: ./
-            Handler: src/error.handler
-            Events:
-                ApiEvent:
+                PutAPI:
                     Type: Api
                     Properties:
-                        Path: /error
-                        Method: GET
-
-    # Lambda function for metrics demonstration
-    MetricsFunction:
-        Type: AWS::Serverless::Function
-        Properties:
-            CodeUri: ./
-            Handler: src/metrics.handler
-            Events:
-                ApiEvent:
-                    Type: Api
-                    Properties:
-                        Path: /metrics
-                        Method: GET
-
-    # SQS queue for event processing demonstration
-    ProcessingQueue:
-        Type: AWS::SQS::Queue
-        Properties:
-            MessageRetentionPeriod: 1209600 # 14 days
-            VisibilityTimeout: 180
-
+                        Path: /
+                        Method: PUT
 Outputs:
-    ApiFunction:
-        Description: 'API Lambda Function ARN'
-        Value: !GetAtt ApiFunction.Arn
+    WebEndpoint:
+        Description: API Gateway endpoint URL for Prod stage
+        Value: !Sub 'https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Prod/'
 
-    ApiEndpoint:
-        Description: 'API Gateway endpoint URL'
-        Value: !Sub 'https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Prod/api/'
-
-    ErrorEndpoint:
-        Description: 'Error demonstration endpoint URL'
-        Value: !Sub 'https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Prod/error/'
-
-    MetricsEndpoint:
-        Description: 'Metrics demonstration endpoint URL'
-        Value: !Sub 'https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Prod/metrics/'
+# More info about Globals: https://github.com/awslabs/serverless-application-model/blob/master/docs/globals.rst
+Globals:
+    Function:
+        LoggingConfig:
+            LogFormat: JSON
+        Runtime: nodejs22.x
+        Architectures:
+            - x86_64
+        MemorySize: 128
+        Timeout: 100
+        Environment:
+            Variables:
+                HIGHLIGHT_PROJECT_ID: '1'
